Navigation improved. Claim and balanceUpdated method addeded
